move some more stuff around
