+++ app/javascript/components/Car/CarForm.test.jsx
@@ -4,7 +4,7 @@ import axios from 'axios';
 import { shallow } from 'enzyme';
 
 import CarForm from './CarForm';
-import wait from '../../libraries/wait';
+import wait from '../../shared/wait';
 
 const props = {
   car: {
@@ -25,23 +25,31 @@ describe(CarForm, () => {
   });
 
   test('it updates the field when edited', () => {
-    const field = wrapper.find('input[type="text"]').first();
-    expect(field.prop('value')).toEqual('BMW');
-    field.simulate('change', {
-      target: { value: 'Toyota' },
-    });
-    expect(wrapper.find('input[type="text"]').first().prop('value')).toEqual('Toyota');
+    const field = () => wrapper.find('input[type="text"]').first();
+
+    expect(field().prop('value')).toEqual('BMW');
+    field().simulate('change', { target: { value: 'Toyota' } });
+    expect(field().prop('value')).toEqual('Toyota');
+  });
+
+  test('fetches all the cars after a successful form submission', async () => {
+    axios.onPut(/cars/).reply(200, { error: null });
+    wrapper.find('form').simulate('submit', { preventDefault() {} });
+    await wait();
+
+    expect(props.fetchCars).toHaveBeenCalled();
   });
 
   test('shows erroneous fields if response has error after form submission', async () => {
-    axios.onPut(/cars/).reply(422, {
-      error: { manufacturer: 'can not be blank' },
-    });
+    const errorField = () => wrapper.find('div.car-edit__error');
+
+    axios.onPut(/cars/).reply(422, { error: { manufacturer: 'can not be blank' } });
+
+    expect(errorField().length).toEqual(0);
 
     wrapper.find('form').simulate('submit', { preventDefault() {} });
     await wait();
-    expect(props.fetchCars).not.toHaveBeenCalled();
 
-    expect(wrapper.find('div.car-edit__error').text()).toEqual('can not be blank');
+    expect(errorField().text()).toEqual('can not be blank');
   });
 });
