+++ app/queries/queries/paginate.rb
@@ -1,7 +1,7 @@
 module Queries
   class Paginate < Queries::Base
     def call
-      @relation.page(@params[:page]).per(10)
+      @relation.page(@params[:page]).per(@params[:per_page] || 10)
     end
   end
 end
