+++ app/queries/queries/search.rb
@@ -1,7 +1,7 @@
 module Queries
   class Search < Queries::Base
     def call
-      return @relation unless @params[:q]
+      return @relation unless @params[:q].present?
 
       @relation.search(@params[:q])
     end
